+++ docs/changelog.md
@@ -30,6 +30,7 @@ Click on the version in the following list to see all changes since the previous
 * refactor: improve readability of mapper for REST representation of a session
 * refactor: rename estimate service in preparation for new custom field logic
 * refactor: rename story point field in preparation for new custom field logic
+* feat: calculation of bounded votes supports non-numeric card values (#102)
 
 ### [4.10.0] - 2020-03-28